+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Release 0.3 (????-??-??)
+* Added game window border and top bar displaying game time, current score and high score
+* Score calculated as snake length
+* Store the best score in a text file
+
 ## Release 0.2 (????-??-??)
 * Added gaps between tiles